+++ src/routes/bug/isdisableddate/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import DateInput from '$lib/DateInput.svelte'
 
-	let min = new Date(2024, 1, 26, 17, 30)
-	let value: Date | undefined
+	let min = $state(new Date(2024, 1, 26, 17, 30))
+	let value: Date | undefined = $state()
 </script>
 
 <DateInput
@@ -14,7 +14,7 @@
 	}}
 />
 <button
-	on:click={() => {
+	onclick={() => {
 		value = new Date(2024, 10, 15)
 	}}>Set to 2024-10-15</button
 >