+++ src/routes/bug/timeprecision/+page.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
 	import DatePicker from '$lib/DatePicker.svelte'
 
-	let value: Date | undefined
+	let value: Date | undefined = $state()
 </script>
 
 <button
-	on:click={() => {
+	onclick={() => {
 		value = new Date(2025, 3, 19, 11, 11, 11, 111)
 	}}>chhh</button
 >